requirements-python.txt
# WaveScope Pipeline Python Requirements
# Enhanced YouTube to Supabase integration and data processing

# YouTube API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# Database
supabase==2.3.0
postgrest==0.13.2

# Environment and Configuration
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# HTTP and Networking
requests==2.31.0
urllib3==2.0.7

# Utilities
certifi==2023.11.17
python-dateutil==2.8.2

# Optional: For advanced analytics (if needed)
# scikit-learn==1.3.2
# scipy==1.11.4